PricePoint price: int -> float

+++ src/PricePoint.py
@@ -1,4 +1,4 @@
 class PricePoint():
-    def __init__(self, timestamp: int, price: int) -> None:
+    def __init__(self, timestamp: int, price: float) -> None:
         self.timestamp = timestamp
         self.price = price
